エンティティのコメント整備

+++ exam-com/src/main/java/com/example/action/model/MBushoDeleteAction.java
@@ -11,21 +11,34 @@ import jp.co.golorp.emarf.exception.OptLockError;
 import jp.co.golorp.emarf.util.Messages;
 import jp.co.golorp.emarf.validation.FormValidator;
 
+/**
+ * 部署マスタ削除
+ *
+ * @author emarfkrow
+ */
 public class MBushoDeleteAction extends BaseAction {
 
-    /**
-     *
-     */
+    /** 部署マスタ削除処理 */
     @Override
     public Map<String, Object> running(final LocalDateTime now, final String id, final Map<String, Object> postJson) {
 
         Map<String, Object> map = new HashMap<String, Object>();
+
+        // 主キーが不足していたらエラー
+        Object bushoId = postJson.get("bushoId");
+        if (bushoId == null) {
+            bushoId = postJson.get("MBusho.bushoId");
+        }
+        if (bushoId == null) {
+            throw new OptLockError("error.cant.delete");
+        }
 
         MBusho e = FormValidator.toBean(MBusho.class.getName(), postJson);
 
         if (e.delete() != 1) {
             throw new OptLockError("error.cant.delete");
         }
+
         map.put("INFO", Messages.get("info.delete"));
 
         return map;
